pyproject.toml
[tool.poetry]
name = "uhi"
version = "0.1.0"
description = ""
authors = ["Andres <842236@unizar.es>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.35.0"
fiona = "^1.9.6"
matplotlib = "^3.9.2"
numpy = "^2.0.1"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
rasterio = "^1.3.10"
requests = "^2.32.3"
scikit-learn = "^1.5.1"
torch = "^2.6.0"
scipy = "^1.15.2"
mkdocs = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
